Fix(ui-client): TTTK-52 Moved UI Client to new file
